requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
black==21.7b0
click==8.0.1
coloredlogs==15.0.1
decorator==5.0.9
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.10.1
greenlet==1.1.0
humanfriendly==9.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
msgpack==1.0.2
mypy-extensions==0.4.3
neovim==0.3.1
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.27
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==1.7.1
pynvim==0.4.3
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==1.1.0
traitlets==5.0.5
wcwidth==0.2.5
yapf==0.31.0
